requirements.txt
streamlit==1.38.0

# LangChain + Extensions
langchain==0.3.1
langchain-community==0.3.1
langchain-chroma==0.1.4
langchain-groq==0.2.0
langchain-openai==0.2.1

# Vector Store + Embeddings
chromadb==0.5.5
sentence-transformers==3.0.1
huggingface-hub==0.24.6
pypdf==5.0.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Utils
python-dotenv==1.0.1
httpx==0.27.0
groq==0.5.0

# Optional (if required anywhere)
Faker==30.0.0
